+++ include/pt/geometry/constant_medium.hpp
@@ -3,6 +3,7 @@
 #include "pt/core/hittable.hpp"
 #include "pt/math/aabb.hpp"
 #include "pt/math/interval.hpp"
+#include "pt/math/scalar.hpp"
 #include <memory>
 
 namespace pt {
@@ -13,7 +14,7 @@ class ray;
 
 class constant_medium final : public hittable {
 public:
-    constant_medium(std::shared_ptr<hittable> boundary, double density, const material* phase_function);
+    constant_medium(std::shared_ptr<hittable> boundary, Float density, const material* phase_function);
 
     [[nodiscard]] aabb bounding_box() const override;
 
@@ -21,7 +22,7 @@ public:
 
 private:
     std::shared_ptr<hittable> boundary;
-    double neg_inv_density{};
+    Float neg_inv_density{};
     const material* phase_function;
 };
