+++ include/pt/geometry/sphere.hpp
@@ -4,6 +4,7 @@
 #include "pt/math/aabb.hpp"
 #include "pt/math/interval.hpp"
 #include "pt/math/ray.hpp"
+#include "pt/math/scalar.hpp"
 #include "pt/math/vec3.hpp"
 
 namespace pt {
@@ -12,9 +13,9 @@ class material;
 
 class sphere final : public hittable {
 public:
-    sphere(const point3& center1, const point3& center2, double radius, const material* mat);
+    sphere(const point3& center1, const point3& center2, Float radius, const material* mat);
 
-    sphere(const point3& static_center, double radius, const material* mat);
+    sphere(const point3& static_center, Float radius, const material* mat);
 
     [[nodiscard]] aabb bounding_box() const override;
 
@@ -22,21 +23,21 @@ public:
 
     [[nodiscard]] vec3 random(const point3& origin) const override;
 
-    [[nodiscard]] double pdf_value(const point3& origin, const vec3& direction) const override;
+    [[nodiscard]] Float pdf_value(const point3& origin, const vec3& direction) const override;
 
 private:
     ray center;
-    double radius;
+    Float radius;
     const material* mat = nullptr;
     aabb bbox;
 
     struct uv_coords {
-        double u{}, v{};
+        Float u{}, v{};
     };
 
     [[nodiscard]] static uv_coords get_sphere_uv(const point3& p);
 
-    [[nodiscard]] static vec3 random_to_sphere(double radius, double distance_squared);
+    [[nodiscard]] static vec3 random_to_sphere(Float radius, Float distance_squared);
 };
 
 } // namespace pt
