+++ include/pt/materials/isotropic.hpp
@@ -2,6 +2,7 @@
 #include "pt/core/hit_record.hpp"
 #include "pt/materials/material.hpp"
 #include "pt/math/ray.hpp"
+#include "pt/math/scalar.hpp"
 #include <optional>
 
 namespace pt {
@@ -14,7 +15,7 @@ public:
 
     [[nodiscard]] std::optional<scatter_record> scatter(const ray& r_in, const hit_record& rec) const override;
 
-    [[nodiscard]] double scattering_pdf(const ray& r_in, const hit_record& rec, const ray& scattered) const override;
+    [[nodiscard]] Float scattering_pdf(const ray& r_in, const hit_record& rec, const ray& scattered) const override;
 
 private:
     const texture* tex;
