+++ include/pt/materials/metal.hpp
@@ -2,6 +2,7 @@
 #include "pt/core/hit_record.hpp"
 #include "pt/materials/material.hpp"
 #include "pt/math/ray.hpp"
+#include "pt/math/scalar.hpp"
 #include "pt/math/vec3.hpp"
 #include <optional>
 
@@ -9,13 +10,13 @@ namespace pt {
 
 class metal final : public material {
 public:
-    explicit metal(const color& albedo, double fuzz);
+    explicit metal(const color& albedo, Float fuzz);
 
     [[nodiscard]] std::optional<scatter_record> scatter(const ray& r_in, const hit_record& rec) const override;
 
 private:
     color albedo;
-    double fuzz{};
+    Float fuzz{};
 };
 
 } // namespace pt
