+++ include/pt/textures/texture.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "pt/math/scalar.hpp"
 #include "pt/math/vec3.hpp"
 
 namespace pt {
@@ -7,7 +8,7 @@ class texture {
 public:
     virtual ~texture();
 
-    [[nodiscard]] virtual color value(double u, double v, const point3& p) const = 0;
+    [[nodiscard]] virtual color value(Float u, Float v, const point3& p) const = 0;
 };
 
 } // namespace pt
